Customer Section Completed(Need to finish account section)

+++ BankingApp/Models/Account.cs
@@ -4,7 +4,7 @@
     {
         public int accountid { get; set; }
         public int userid { get; set; }
-        public int balance { get; set; }
+        public decimal balance { get; set; }
         public string accounttype { get; set; }
     }
 }
